+++ app/src/main/java/io/github/wiiznokes/gitnote/ui/viewmodel/GridViewModel.kt
@@ -18,7 +18,6 @@ import io.github.wiiznokes.gitnote.helper.FrontmatterParser
 import io.github.wiiznokes.gitnote.helper.NameValidation
 import io.github.wiiznokes.gitnote.manager.StorageManager
 import io.github.wiiznokes.gitnote.ui.model.FileExtension
-import io.github.wiiznokes.gitnote.ui.model.FolderDisplayMode
 import io.github.wiiznokes.gitnote.ui.model.GridNote
 import io.github.wiiznokes.gitnote.ui.model.NoteViewType
 import io.github.wiiznokes.gitnote.ui.model.SortOrder
@@ -268,7 +267,7 @@ class GridViewModel : ViewModel() {
 
 
     @OptIn(ExperimentalCoroutinesApi::class)
-    val pagingFlow = prefs.folderDisplayMode.getFlow().flatMapLatest { folderDisplayMode ->
+    val pagingFlow = prefs.includeSubfolders.getFlow().flatMapLatest { includeSubfolders ->
         combine(
             currentNoteFolderRelativePath,
             prefs.sortOrder.getFlow(),
@@ -280,9 +279,9 @@ class GridViewModel : ViewModel() {
                 config = PagingConfig(pageSize = 50),
                 pagingSourceFactory = {
                     if (query.isEmpty()) {
-                        dao.gridNotes(currentNoteFolderRelativePath, sortOrder, folderDisplayMode, selectedTag)
+                        dao.gridNotes(currentNoteFolderRelativePath, sortOrder, includeSubfolders, selectedTag)
                     } else {
-                        dao.gridNotesWithQuery(currentNoteFolderRelativePath, sortOrder, query, folderDisplayMode, selectedTag)
+                        dao.gridNotesWithQuery(currentNoteFolderRelativePath, sortOrder, query, includeSubfolders, selectedTag)
                     }
                 }
             ).flow.cachedIn(viewModelScope)